refactor: remove spaces from classname
