+++ src/components/InputBox/index.tsx
@@ -2,8 +2,8 @@ import React, { SetStateAction, useState } from 'react';
 
 import Button from '@/components/Button';
 
-import EyeIcon from '../../../public/assets/icons/eye';
-import EyeCloseIcon from '../../../public/assets/icons/eyeClose';
+import EyeIcon from '../icons/eye';
+import EyeCloseIcon from '../icons/eyeClose';
 
 type Props = {
     type: string;
