+++ src/components/Navbar/index.tsx
@@ -5,8 +5,8 @@ import Button from '@/components/Button';
 import { TINY_API_GOOGLE_LOGIN, TINY_API_LOGOUT } from '@/constants/url';
 import IsAuthenticated from '@/hooks/isAuthenticated';
 
-import DownArrowIcon from '../../../public/assets/icons/downArrow';
-import GoogleIcon from '../../../public/assets/icons/google';
+import DownArrowIcon from '../icons/downArrow';
+import GoogleIcon from '../icons/google';
 import ProfileIcon from '../ProfileIcon/ProfileIcon';
 
 const Navbar: React.FC = () => {
